feat: renamed folders
