+++ src/constants/initialValues.ts
@@ -1,0 +1,24 @@
+import { FormValues } from "../types/types";
+
+export const initialValues: FormValues = {
+  lastName: '',
+  firstName: '',
+  middleName: '',
+  middleNameIsActive: false,
+  rnokpp: '',
+  rnokppIsActive: false,
+  birthDate: '',
+  gender: '',
+  birthCountry: '',
+  birthPlace: '',
+  secretWord: '',
+  email: '',
+  phone: '',
+  contactMethod: '',
+  documentType: '',
+  documentSeriesNumber: '',
+  issuedBy: '',
+  issuedDate: '',
+  expiryDate: '',
+  registryNumber: ''
+}
